@@ -1,11 +1,13 @@
 <?php
 namespace Aura\Uri;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Test class for Path.
  * Generated by PHPUnit on 2012-07-21 at 15:45:14.
  */
-class PathTest extends \PHPUnit_Framework_TestCase
+class PathTest extends TestCase
 {
     /**
      * @var Path
@@ -16,7 +18,7 @@ class PathTest extends \PHPUnit_Framework_TestCase
      * Sets up the fixture, for example, opens a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    public function setUp() : void
     {
         parent::setUp();
         $this->path = new Path;
@@ -26,13 +28,13 @@ class PathTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, closes a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    public function tearDown() : void
     {
         parent::tearDown();
     }
 
     /**
-     * @covers Aura\Uri\Path::__toString
+     * @covers \Aura\Uri\Path::__toString
      */
     public function test__toString()
     {
@@ -43,7 +45,7 @@ class PathTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Path::setFromString
+     * @covers \Aura\Uri\Path::setFromString
      */
     public function testSetFromString()
     {
@@ -59,8 +61,8 @@ class PathTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Path::setFormat
-     * @covers Aura\Uri\Path::getFormat
+     * @covers \Aura\Uri\Path::setFormat
+     * @covers \Aura\Uri\Path::getFormat
      */
     public function testSetAndGetFormat()
     {
