@@ -1,11 +1,13 @@
 <?php
 namespace Aura\Uri;
 
+use PHPUnit\Framework\TestCase;
+
 /**
  * Test class for Query.
  * Generated by PHPUnit on 2012-07-21 at 15:45:19.
  */
-class QueryTest extends \PHPUnit_Framework_TestCase
+class QueryTest extends TestCase
 {
     /**
      * @var Query
@@ -16,7 +18,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
      * Sets up the fixture, for example, opens a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    public function setUp() : void
     {
         parent::setUp();
         $this->query = new Query;
@@ -26,13 +28,13 @@ class QueryTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, closes a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    public function tearDown() : void
     {
         parent::tearDown();
     }
 
     /**
-     * @covers Aura\Uri\Query::__toString
+     * @covers \Aura\Uri\Query::__toString
      */
     public function test__toString()
     {
@@ -43,7 +45,7 @@ class QueryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Query::setFromString
+     * @covers \Aura\Uri\Query::setFromString
      */
     public function testSetFromString()
     {
