@@ -1,13 +1,14 @@
 <?php
 namespace Aura\Uri\Url;
 
+use PHPUnit\Framework\TestCase;
 use Aura\Uri\PublicSuffixList;
 
 /**
  * Test class for Factory.
  * Generated by PHPUnit on 2012-07-21 at 17:13:55.
  */
-class FactoryTest extends \PHPUnit_Framework_TestCase
+class FactoryTest extends TestCase
 {
     protected function newFactory($server = [])
     {
@@ -16,7 +17,7 @@ class FactoryTest extends \PHPUnit_Framework_TestCase
     }
     
     /**
-     * @covers Aura\Uri\Url\Factory::newInstance
+     * @covers \Aura\Uri\Url\Factory::newInstance
      */
     public function testNewInstance()
     {
@@ -27,7 +28,7 @@ class FactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url\Factory::newCurrent
+     * @covers \Aura\Uri\Url\Factory::newCurrent
      */
     public function testNewCurrent()
     {
