@@ -1,13 +1,14 @@
 <?php
 namespace Aura\Uri;
 
+use PHPUnit\Framework\TestCase;
 use Aura\Uri\Url\Factory as UrlFactory;
 
 /**
  * Test class for Url.
  * Generated by PHPUnit on 2012-07-21 at 15:46:30.
  */
-class UrlTest extends \PHPUnit_Framework_TestCase
+class UrlTest extends TestCase
 {
     /**
      * @var Url
@@ -28,7 +29,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
      * Sets up the fixture, for example, opens a network connection.
      * This method is called before a test is executed.
      */
-    protected function setUp()
+    public function setUp() : void
     {
         parent::setUp();
         $file = dirname(dirname(dirname(__DIR__))) . DIRECTORY_SEPARATOR
@@ -43,7 +44,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
      * Tears down the fixture, for example, closes a network connection.
      * This method is called after a test is executed.
      */
-    protected function tearDown()
+    public function tearDown() : void
     {
         parent::tearDown();
     }
@@ -72,7 +73,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::__toString
+     * @covers \Aura\Uri\Url::__toString
      */
     public function test__toString()
     {
@@ -81,7 +82,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::__get
+     * @covers \Aura\Uri\Url::__get
      */
     public function test__get()
     {
@@ -100,7 +101,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::get
+     * @covers \Aura\Uri\Url::get
      */
     public function testGet()
     {
@@ -110,7 +111,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::getFull
+     * @covers \Aura\Uri\Url::getFull
      */
     public function testGetFull()
     {
@@ -119,7 +120,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::getSchemeless
+     * @covers \Aura\Uri\Url::getSchemeless
      */
     public function testGetSchemeless()
     {
@@ -129,7 +130,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setScheme
+     * @covers \Aura\Uri\Url::setScheme
      */
     public function testSetScheme()
     {
@@ -139,7 +140,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setUser
+     * @covers \Aura\Uri\Url::setUser
      */
     public function testSetUser()
     {
@@ -149,7 +150,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setPass
+     * @covers \Aura\Uri\Url::setPass
      */
     public function testSetPass()
     {
@@ -159,7 +160,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setHost
+     * @covers \Aura\Uri\Url::setHost
      */
     public function testSetHost()
     {
@@ -169,7 +170,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setPort
+     * @covers \Aura\Uri\Url::setPort
      */
     public function testSetPort()
     {
@@ -179,7 +180,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setPath
+     * @covers \Aura\Uri\Url::setPath
      */
     public function testSetPath()
     {
@@ -189,7 +190,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setQuery
+     * @covers \Aura\Uri\Url::setQuery
      */
     public function testSetQuery()
     {
@@ -199,7 +200,7 @@ class UrlTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers Aura\Uri\Url::setFragment
+     * @covers \Aura\Uri\Url::setFragment
      */
     public function testSetFragment()
     {
